agrilabo/requirements.txt
asgiref==3.6.0
autopep8==2.0.2
certifi==2022.12.7
charset-normalizer==3.1.0
cloudinary==1.32.0
distlib==0.3.6
dj-database-url==1.3.0
Django==4.2
django-cloudinary-storage==0.3.0
django-environ==0.10.0
filelock==3.11.0
gunicorn==20.1.0
idna==3.4
mysqlclient==2.1.1
Pillow==9.5.0
platformdirs==3.2.0
psycopg2-binary==2.9.6
pycodestyle==2.10.0
requests==2.28.2
six==1.16.0
sqlparse==0.4.3
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.15
whitenoise==6.4.0
